+++ libft/ft_strlcat.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mzoheir <mzoheir@student.42.fr>            +#+  +:+       +#+        */
+/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 13:53:25 by mzoheir           #+#    #+#             */
-/*   Updated: 2023/02/07 18:59:34 by mzoheir          ###   ########.fr       */
+/*   Updated: 2023/08/08 23:23:46 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
+int	ft_strlcat(char *dst, const char *src, int dstsize)
 {
-	size_t	i;
-	size_t	ldst;
-	size_t	lsrc;
+	int	i;
+	int	ldst;
+	int	lsrc;
 
 	if (dstsize == 0 && dst == 0)
 		return (0);
